+++ src/Item/EditItem/EditItem.test.js
@@ -9,8 +9,9 @@ import {
   useInstanceQuery,
   useHolding,
 } from '../../common/hooks';
-import { useItem } from '../hooks';
+import { useItem, useItemMutation } from '../hooks';
 import EditItem from './EditItem';
+import ItemForm from '../../edit/items/ItemForm';
 
 jest.mock('../../edit/items/ItemForm', () => jest.fn().mockReturnValue('ItemForm'));
 jest.mock('../../hooks/useCallout', () => jest.fn().mockReturnValue({ sendCallout: jest.fn() }));
@@ -22,6 +23,7 @@ jest.mock('../../common/hooks', () => ({
 jest.mock('../hooks', () => ({
   ...jest.requireActual('../hooks'),
   useItem: jest.fn().mockReturnValue({ item: {}, isLoading: false }),
+  useItemMutation: jest.fn().mockReturnValue({ mutateItem: jest.fn() }),
 }));
 
 const defaultProps = {
@@ -51,10 +53,13 @@ const renderEditItem = (props = {}) => render(
 
 
 describe('EditItem', () => {
+  const mutateItem = jest.fn(() => Promise.resolve({ data: {} }));
+
   beforeEach(() => {
     useInstanceQuery.mockClear();
     useHolding.mockClear();
     useItem.mockClear();
+    useItemMutation.mockClear().mockReturnValue({ mutateItem });
   });
 
   it('should render ItemForm', () => {
@@ -69,5 +74,29 @@ describe('EditItem', () => {
     renderEditItem();
 
     expect(screen.getByText('LoadingView')).toBeInTheDocument();
+  });
+
+
+  it('should render LoadingView if page is loading', () => {
+    useInstanceQuery.mockReturnValue({ isLoading: true });
+
+    renderEditItem();
+
+    expect(screen.getByText('LoadingView')).toBeInTheDocument();
+  });
+
+  it('should handle optimistic locking exception', () => {
+    const error = new Error({ response: 'optimistic locking' });
+    mutateItem.mockClear().mockImplementation(() => Promise.reject(error));
+
+    renderEditItem();
+
+    ItemForm.mock.calls[0][0].onSubmit({
+      id: 'itemId',
+      permanentLocationId: '',
+      temporaryLocationId: '',
+    });
+
+    expect(mutateItem).toHaveBeenCalled();
   });
 });